requirements.txt
greenlet==1.0.0
Flask_Cors==3.0.10
Flask_MySQLdb==0.2.0
Flask_JWT_Extended==4.2.1
ordereddict==1.1
PyYAML==5.4.1
gunicorn==20.1.0
pip==21.1.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
Jinja2==2.11.3
click==7.1.2
python-dotenv==0.17.1
itsdangerous==1.1.0
visitor==0.1.3
mysqlclient==2.0.3
six==1.15.0
Flask==1.1.2
setuptools==49.2.1
PyJWT==2.1.0